# (docker) fix tag initialization, optimize loading