--- a/app/scripts/modules/netflix/fastProperties/domain/scope.domain.spec.ts
+++ b/app/scripts/modules/netflix/fastProperties/domain/scope.domain.spec.ts
@@ -25,7 +25,7 @@ describe('Scope Domain Spec', function () {
         expect(result[prop]).toBe(platformProperty[prop]);
       });
 
-      expect(result['propertyId']).toBeUndefined()
+      expect(result['propertyId']).toBeUndefined();
 
     });
   });