This was initially a one-line change: moving `this.updateOrganizationsList()` down so it executes after `this.updateRepositoryMap(images)`. This ensures tags are pre-populated when the controller initializes with existing data.

All the other changes are performance optimizations. I wouldn't have even noticed them, but I had my browser console open and noticed, when the modal opens, warnings like this:
```
[Violation] Long running JavaScript task took 974ms
```
It was pretty consistently running between 900 and 1100 ms, and it's blocking. Running the profiler showed almost all of that time being spent in the `all.includes(...)` calls - not what I expected, but I guess it makes sense when there are 35k+ images. 

The easy fix is to jam everything into the array, then remove duplicates at the end via `_.uniq`. This drops initialization to < 150ms (I didn't measure, but Chrome won't warn unless it's > 150ms), and the select fields are available almost immediately, so it's a nice little fix.

Added a couple of very rudimentary tests to verify de-duping works as expected.

I manually tested against all the Titus pieces and expect the other usages will work fine, too.

@icfantv or @danielpeach PTAL