@icfantv 2017-02-07
LGTM.  Nice work.  I'll defer to @danielpeach to confirm it still works fine w/ GCP.